# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FieldForce Live — Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users for the admin's customer
      parameters:
        - name: role
          in: query
          schema:
            $ref: "#/components/schemas/UserRole"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/UserStatus"
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a user or admin (geocodes addresses, generates onboarding invite for USER role)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreationResult"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user's profile + addresses + marked places
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user's profile fields
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{id}/onboarding-invite:
    get:
      operationId: getOnboardingInvite
      tags: [onboarding]
      summary: Fetch (or lazily create) the onboarding invite + shareable link for a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingInvite"
  /users/{id}/emergency-alert:
    post:
      operationId: triggerEmergencyAlert
      tags: [alerts]
      summary: Trigger an emergency "call back now" alert for a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyAlertInput"
      responses:
        "201":
          description: Alert triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"
  /users/{id}/alerts:
    get:
      operationId: listUserAlerts
      tags: [alerts]
      summary: List emergency alerts for a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyAlert"

  # ---------- Marked places ----------
  /users/{id}/marked-places:
    get:
      operationId: listMarkedPlaces
      tags: [places]
      summary: List a user's marked places (up to 5)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Marked places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarkedPlace"
    post:
      operationId: createMarkedPlace
      tags: [places]
      summary: Add a marked place for a user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkedPlaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkedPlace"
  /marked-places/{placeId}:
    delete:
      operationId: deleteMarkedPlace
      tags: [places]
      summary: Remove a marked place
      parameters:
        - name: placeId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ---------- Live map + dashboard ----------
...
  /onboarding/{token}:
    get:
      operationId: getOnboardingByToken
      tags: [onboarding]
      summary: Resolve an onboarding token to its landing-page data
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Onboarding info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingLanding"
        "404":
          description: Invalid token
...
components:
  schemas:
...
    UserRole:
      type: string
      enum: [ADMIN, USER]
    UserStatus:
      type: string
      enum: [INVITED, ACTIVE, SUSPENDED]
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
    AddressType:
      type: string
      enum: [OFFICE, BASE_OFFICE, SITE_OFFICE, HOME]
...
    OnboardingChannel:
      type: string
      enum: [WHATSAPP, EMAIL]
...
    Address:
      type: object
      required: [type, rawAddress, latitude, longitude]
      properties:
        type:
          $ref: "#/components/schemas/AddressType"
        rawAddress: { type: string }
        latitude: { type: number }
        longitude: { type: number }
    AddressInput:
      type: object
      required: [type, rawAddress]
      properties:
        type:
          $ref: "#/components/schemas/AddressType"
        rawAddress: { type: string, minLength: 1 }
    MarkedPlace:
      type: object
      required: [id, userId, label, latitude, longitude]
      properties:
        id: { type: integer }
        userId: { type: integer }
        label: { type: string }
        latitude: { type: number }
        longitude: { type: number }
    MarkedPlaceInput:
      type: object
      required: [label, rawAddress]
      properties:
        label: { type: string, minLength: 1 }
        rawAddress: { type: string, minLength: 1 }
    User:
      type: object
      required:
        [
          id,
          customerId,
          firstName,
          lastName,
          gender,
          employeeCode,
          phoneNumber,
          email,
          role,
          status,
          consentGivenAt,
          createdAt,
        ]
      properties:
        id: { type: integer }
        customerId: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        gender:
          $ref: "#/components/schemas/Gender"
        employeeCode: { type: string }
        phoneNumber: { type: string }
        email: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          $ref: "#/components/schemas/UserStatus"
        consentGivenAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [addresses, markedPlaces]
          properties:
            addresses:
              type: array
              items:
                $ref: "#/components/schemas/Address"
            markedPlaces:
              type: array
              items:
                $ref: "#/components/schemas/MarkedPlace"
    UserInput:
      type: object
      required:
        [firstName, lastName, gender, employeeCode, phoneNumber, email, role, addresses]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        gender:
          $ref: "#/components/schemas/Gender"
        employeeCode: { type: string, minLength: 1 }
        phoneNumber: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        username: { type: string }
        role:
          $ref: "#/components/schemas/UserRole"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/AddressInput"
        markedPlaces:
          type: array
          items:
            $ref: "#/components/schemas/MarkedPlaceInput"
    UserUpdate:
      type: object
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        phoneNumber: { type: string, minLength: 1 }
        email: { type: string, minLength: 1 }
        status:
          $ref: "#/components/schemas/UserStatus"
    UserCreationResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        onboardingInvite:
          $ref: "#/components/schemas/OnboardingInvite"
    OnboardingInvite:
      type: object
      required: [userId, token, channel, deepLink, createdAt]
      properties:
        userId: { type: integer }
        token: { type: string }
        channel:
          $ref: "#/components/schemas/OnboardingChannel"
        deepLink: { type: string }
        createdAt: { type: string, format: date-time }
        usedAt: { type: ["string", "null"], format: date-time }
    OnboardingLanding:
      type: object
      required: [userFirstName, status, consentGivenAt, iosStoreUrl, androidStoreUrl, huaweiStoreUrl]
      properties:
        userFirstName: { type: string }
        status:
          $ref: "#/components/schemas/UserStatus"
        consentGivenAt: { type: ["string", "null"], format: date-time }
        iosStoreUrl: { type: string }
        androidStoreUrl: { type: string }
        huaweiStoreUrl: { type: string }
    EmergencyAlertInput:
      type: object
      properties:
        message: { type: string }
    EmergencyAlert:
      type: object
      required: [id, userId, message, triggeredAt, acknowledgedAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        triggeredByAdminId: { type: ["integer", "null"] }
        message: { type: string }
        triggeredAt: { type: string, format: date-time }
        acknowledgedAt: { type: ["string", "null"], format: date-time }
        direction: { type: string, enum: [ADMIN_TO_USER, USER_TO_ADMIN] }
        lat: { type: ["number", "null"] }
        lng: { type: ["number", "null"] }
...